extract client and server program behaviour from main package
